# openapi.yaml
openapi: 3.1.0
info:
  title: Scrutineer Skill API
  version: 0.3.0
  description: |
    A narrow HTTP surface that claude-code skills use while they run. The
    goal is enough to let a skill orchestrate other skills and read prior
    scan outputs without pushing that logic into scrutineer's Go code.

    Every scrutineer skill scan is issued a bearer token that is staged into
    the workspace context.json. Skills authenticate with that token. A token
    is only valid while its originating scan is running, and only grants
    access to the repository the scan was issued against.

    The `/v1/*` paths are an unauthenticated, localhost-only export surface
    intended for external scripts (jq, cron, CI dumps). They emit JSONL and
    rely on the same host-header check as the browser UI; see threatmodel.md.
servers:
  - url: http://127.0.0.1:8080/api
    description: Local scrutineer server
security:
  - bearerAuth: []
paths:
  /repositories/{id}:
    get:
      summary: Fetch a repository summary
      operationId: getRepository
      parameters:
        - $ref: '#/components/parameters/RepositoryID'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Repository' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /repositories/{id}/scans:
    get:
      summary: List scans for a repository
      operationId: listRepositoryScans
      parameters:
        - $ref: '#/components/parameters/RepositoryID'
        - name: status
          in: query
          schema:
            type: string
            enum: [queued, running, done, failed]
        - name: skill
          in: query
          schema: { type: string }
          description: Filter by skill name (matches Scan.skill_name).
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Scan' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /repositories/{id}/maintainers:
    get:
      summary: List maintainers linked to this repository
      operationId: listRepositoryMaintainers
      parameters:
        - $ref: '#/components/parameters/RepositoryID'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Maintainer' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /repositories/{id}/packages:
    get:
      summary: List published packages produced by this repository
      operationId: listRepositoryPackages
      parameters:
        - $ref: '#/components/parameters/RepositoryID'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Package' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /repositories/{id}/advisories:
    get:
      summary: List published advisories affecting this repository's packages
      operationId: listRepositoryAdvisories
      parameters:
        - $ref: '#/components/parameters/RepositoryID'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Advisory' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /repositories/{id}/dependents:
    get:
      summary: List downstream dependents of this repository's packages
      operationId: listRepositoryDependents
      parameters:
        - $ref: '#/components/parameters/RepositoryID'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Dependent' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /repositories/{id}/dependencies:
    get:
      summary: List dependencies declared by this repository's manifests
      operationId: listRepositoryDependencies
      parameters:
        - $ref: '#/components/parameters/RepositoryID'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Dependency' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /repositories/{id}/findings:
    get:
      summary: List findings across every scan on this repository
      operationId: listRepositoryFindings
      parameters:
        - $ref: '#/components/parameters/RepositoryID'
        - name: severity
          in: query
          schema: { type: string, enum: [Critical, High, Medium, Low] }
        - name: status
          in: query
          schema:
            type: string
            enum: [new, enriched, triaged, ready, reported, acknowledged, fixed, published, rejected, duplicate]
        - name: skill
          in: query
          description: Only return findings produced by scans of this skill (e.g. `semgrep`, `security-deep-dive`).
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/FindingSummary' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /repositories/{id}/dependency-findings:
    get:
      summary: List findings on libraries this repository depends on
      description: >
        Joins this repository's Dependency rows against every Package row
        scrutineer knows about and returns the live findings on the matched
        library repositories. git-pkgs ecosystem names (gem, golang,
        composer) are normalised to packages.ecosyste.ms names (rubygems,
        go, packagist) before matching. Findings already marked fixed,
        rejected or duplicate are excluded. Consumed by the reachability
        skill.
      operationId: listRepositoryDependencyFindings
      parameters:
        - $ref: '#/components/parameters/RepositoryID'
        - name: severity
          in: query
          schema: { type: string, enum: [Critical, High, Medium, Low] }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/DependencyFinding' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /repositories/{id}/skills/{name}/run:
    post:
      summary: Enqueue a skill scan against this repository
      operationId: runSkill
      parameters:
        - $ref: '#/components/parameters/RepositoryID'
        - name: name
          in: path
          required: true
          schema: { type: string }
          description: Skill name (matches the SKILL.md `name` field).
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                model:
                  type: string
                  description: Override model; defaults to the system default.
                ref:
                  type: string
                  description: Git ref to checkout; empty means the repository's default branch.
                profile:
                  type: string
                  description: |
                    Runner profile name (e.g. `php`). Empty auto-detects from the clone via `brief`;
                    `default` forces the default runner image. Unknown names return 400, as do
                    overrides that conflict with the skill's `requires_profile` constraint.
      responses:
        '201':
          description: Scan enqueued
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Scan' }
        '400':
          description: Bad request (unknown profile or profile mismatch with the skill's `requires_profile`)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404':
          description: Skill or repository not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
  /findings/{id}/notes:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: integer, format: int64 }
    get:
      summary: List notes on a finding
      operationId: listFindingNotes
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/FindingNote' }
    post:
      summary: Append a note to a finding
      operationId: addFindingNote
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [body]
              properties:
                body: { type: string }
                by:   { type: string }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FindingNote' }
  /findings/{id}/reviews:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: integer, format: int64 }
    get:
      summary: List structured human verdicts on a finding
      operationId: listFindingReviews
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/FindingReview' }
    post:
      summary: Record one structured human verdict on a finding
      description: |
        Verdicts mirror the revalidate skill's enum so reviewer agreement
        with the model is computed by direct string comparison. When
        automated_outcome is omitted the server snapshots the latest
        revalidate verdict from the finding's notes; explicit values
        win to support older notes the operator wants to re-pin.
      operationId: addFindingReview
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [verdict]
              properties:
                verdict:
                  type: string
                  enum: [true_positive, false_positive, already_fixed, uncertain]
                reason: { type: string }
                reviewer: { type: string }
                automated_outcome:
                  type: string
                  description: Optional override for the automation outcome the human is judging. When empty the server pulls it from the finding's latest revalidate note.
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FindingReview' }
...
  /findings/{id}/communications:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: integer, format: int64 }
    get:
      summary: List logged communications for a finding
      operationId: listFindingCommunications
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/FindingCommunication' }
    post:
      summary: Log a communication (sent or received) about a finding
      operationId: addFindingCommunication
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FindingCommunication' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FindingCommunication' }
  /findings/{id}/references:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: integer, format: int64 }
    get:
      summary: List references attached to a finding
      operationId: listFindingReferences
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/FindingReference' }
    post:
      summary: Attach an external reference to a finding
      operationId: addFindingReference
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FindingReference' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FindingReference' }
  /findings/{id}/labels:
    put:
      summary: Replace the set of labels on a finding
      operationId: setFindingLabels
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [labels]
              properties:
                labels:
                  type: array
                  items: { type: string }
                  description: Label names. Missing labels are created on demand.
      responses:
        '204': { description: Updated }
  /findings/{id}/history:
    get:
      summary: List the history rows recorded for a finding
      operationId: listFindingHistory
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/FindingHistory' }
  /findings/{id}/skills/{name}/run:
    post:
      summary: Enqueue a finding-scoped skill (verify, revalidate, patch, disclose)
      operationId: runFindingSkill
      description: |
        The authenticated scan's repository must own the finding. The scan
        that gets enqueued has FindingID set so its skill body can read
        `scrutineer.finding_id` from context.json and act on one finding
        specifically.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
          description: Finding id.
        - name: name
          in: path
          required: true
          schema: { type: string }
          description: Skill name.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                model: { type: string }
      responses:
        '201':
          description: Scan enqueued
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Scan' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404':
          description: Finding or skill not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /scans/{id}:
    get:
      summary: Fetch a scan with its report (if complete)
      operationId: getScan
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ScanDetail' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /scans/{id}/validate-report:
    post:
      summary: Validate a candidate report.json against the skill's schema
      description: >
        Lets a running skill check its report against its skill's schema using
        scrutineer's own validator, instead of installing a JSON Schema library
        inside the runner container. The request body is the candidate report
        JSON; a 200 response carries {"valid": true} when it conforms, or
        {"valid": false, "errors": "..."} listing the failures. A skill with no
        schema returns {"valid": true, "note": "skill has no schema"}. The token
        may only validate the scan it was issued for. Body is capped at 1 MB.
      operationId: validateReport
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, additionalProperties: true }
      responses:
        '200':
          description: Validation result (valid true or false; never a schema error)
          content:
            application/json:
              schema:
                type: object
                properties:
                  valid: { type: boolean }
                  errors: { type: string }
                  note: { type: string }
                required: [valid]
        '400':
          description: Scan has no associated skill
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '413':
          description: Report body exceeds the 1 MB limit
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /findings/{id}:
    get:
      summary: Fetch one finding with its six-step prose
      operationId: getFinding
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FindingDetail' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /skills:
    get:
      summary: List installed skills
      operationId: listSkills
      parameters:
        - name: active
          in: query
          schema: { type: boolean }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/SkillSummary' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    RepositoryID:
      name: id
      in: path
      required: true
      schema: { type: integer, format: int64 }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
    Repository:
      type: object
      required: [id, url, name]
      properties:
        id:             { type: integer, format: int64 }
        url:            { type: string, format: uri }
        name:           { type: string }
        full_name:      { type: string }
        default_branch: { type: string }
        html_url:       { type: string }
        stars:          { type: integer }
        forks:          { type: integer }
        archived:       { type: boolean }
        languages:      { type: string, description: Comma-joined language names. }
        license:        { type: string }
        fork:           { type: string, description: 'owner/name of the staging fork in fork_org, when one exists.' }
        posture:        { type: string, enum: [ready, partial, unprepared], description: 'Disclosure-readiness tier from the posture skill. Empty when posture has not run.' }
        posture_summary: { type: string }
...